Add the missing content_add_metadata() method to the buffer proxy storage
This is especially necessary to make the graph replayer for content
objects work ok, because we currently have many batches of kafka
messages in which there are several occurrences of the same message,
making the storage used by the replayer drop these for false hash
collisions.

Having the buffer proxy implement this method allows to use it as a
deduplication layer (in addition to its buffer main purpose) which is
needed (until this is made less fragile) for the graph replayer
processing content a but more reliably.