Make the tenacious proxy able to report errors in redis
If a (redis) reporter is configured in a replayer session, use it in the
tenacious proxy storage (if any) to report objects that failed to be
ingested by the tenacious proxy in its backend storage.

Usually these are caught by the validation process in
ModelObjectDeserilaizer but this may not always be the case.

The error "schema" is not the same as the one for the deserializer
because we do not have access to the kafka message at this point, plus
using the object key as redis key has proven to be not ideal, since
several objects (are several attempts for the same object) may be
replicated but only one of them will be kept by redis.

Here we use a "<timestamp_as_iso_date>/<object_type>" as key. It's not
100% guaranteed not to collide but should good enough.

The stored value is a yaml-encoded dict with 2 keys: obj, the dict
representation of the object (ModelObject.to_dict()) and exc, the
original exception dumped as a list of str (as provided by
traceback.format_exception).

Whether it would be wise to update other error reporting to adapt a
similar schema is debatable.