Improve the replayer error reporting format in redis
Following the previous commit, update the existing replayer error reporting
logic to match the tenacious proxy format.

Keys are now prefixed with a timestamp (iso format), eg.
`snapshot:0123...` -> `2025-09-16T09:55:19.373757+00:00/snapshot:0123...`

Values used to be the kafka message only, it now replaced with the yaml
representation of a dict which keys are `kafka_message` (the kafka
message that failed to be converted as a ModelObject and inserted in the
storage), `obj` (dict representation of the ModelObject, resulting from
calling its `.to_dict()` method) and `exc` (list of strings representation
of the exception, as produces by `traceback.format_exception` or None if
no exception was raised).

Also remove the usage of the swhid as key in favor of the
`<object_type>:<hexid>` representation, for the sake of consistency
(since not all ModelObject have a swhid representation).