cassandra: Update revision_log implementation into an iterative one
With the previous recursive implementation to fetch revision log, it
was easy to trigger an error by setting the limit value to 1000
which is the default Python recursion limit.

Remove use of recursion and use iterative approach instead to perform
a DFS on the revisions history.

It should also speedup that operation a bit with the cassanndra backend.