tests: Fix tests collection errors when using pytest-xdist
When using pytest-xdist it is not possible to have tests that differ
in order or their amount across workers.

This is especially true with pytest.mark.parametrize, when values are
produced with sets or other unordered iterables/generators.

Ensure to sort the DEFAULT_ALGORITHMS set when used as tests parameter
to workaround that issue.